requirements.txt
google-adk
google-genai
requests
browser-use-sdk
python-dotenv
youtube-search-python
moviepy
Pillow
imageio-ffmpeg
pyttsx3
Flask
flask-cors
redis
safetensors
scipy
scikit-image
numpy<2
kornia
facexlib
yacs
gfpgan
torchvision
torch==2.2.2
torchvision==0.17.2
librosa
opencv-python
pydub
browser_use_sdk
